Add CallbackNodeTrait
